+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - include USWDS 3.0 for new apps
+- use postcss-minify instead of the sass gem for CSS minimization
 
 ## [0.5.3] - 2022-06-06